# internal/handler/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Host API
  description: |
    Internal static website hosting. Every site is served at the root of its
    own host, `<sitename>.<owner>.<base>/` (for example
    `todo.alice.example.com/`). The owner host `<owner>.<base>/` serves the
    person's or team's index page. Until the owner's certificate is ready it
    also serves their sites at `<owner>.<base>/<sitename>/`; after that,
    those addresses, and `/api/sites/<sitename>/...` on the owner host,
    redirect to each site's own host with path and query kept (`301` for GET
    and HEAD, `308` otherwise). v1.2 addresses of the form
    `<owner>--<sitename>.<base>` only redirect. Always use the `url` from the
    latest response rather than composing an address. Uploads may
    be gzip-compressed tar or ZIP archives; nested regular-file downloads such
    as ZIP and DMG artifacts are supported.

    A site is owned by a person or a team. Team members change a team's
    sites through owner-qualified collaboration routes with optimistic
    `ETag` / `If-Match` protection.

    Every site has an access level: `only_me` (the owner, or the team's
    members; the default for a new site), `specific` (plus named viewers),
    `company` (anyone signed in with the link),
    `listed` (company, plus showcase and search), or `network` (anyone who
    can reach the server, no sign-in). `network` is requested with a reason
...
    network site's pages, assets and state, and change nothing.

...

...

...

...
servers:
...
paths:
...
  /api/sites:
    get:
      summary: List sites
      description: Returns the authenticated user's sites with daily server-side analytics. Admins see all sites.
      tags: [Sites]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
      responses:
        '200':
          description: List of sites
          headers:
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Site'
        '400':
          $ref: '#/components/responses/SkillVersionRequired'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/sites/{sitename}:
    post:
      summary: Create a new site
      description: |
        Upload a gzip-compressed tar or ZIP archive. The compressed request is
        limited to 100 MiB; each regular file and all files in aggregate are
        limited to 500 MiB of bytes actually read; at most 50,000 entries are
        accepted. Complete gzip/TAR output is independently limited to 768 MiB,
        including framing, extended metadata, padding, and trailing bytes;
        unique path-tracking metadata is limited to 64 MiB. Entries must be
        regular files or directories with canonical, confined paths. Symlinks
        and special files are rejected.

        Extensions use a denylist, not an allowlist: source scripts (`.sh`,
        `.bash`, `.zsh`, `.fish`, `.bat`, `.cmd`, `.ps1`, `.py`, `.pyc`, `.rb`,
        `.pl`, `.go`, `.php`) and executables, installers, packages and disk
        images (`.exe`, `.dll`, `.msi`, `.msix`, `.appx`, `.scr`, `.com`,
        `.pif`, `.cpl`, `.hta`, `.vbs`, `.vbe`, `.jse`, `.wsf`, `.wsh`, `.lnk`,
        `.reg`, `.jar`, `.apk`, `.aab`, `.pkg`, `.deb`, `.rpm`, `.iso`, `.img`)
        are rejected case-insensitively. Other regular files,
        including nested ZIP and DMG downloads, are supported, except entries
        whose basename is `.DS_Store` or starts with `._`; those macOS metadata
        files are omitted. The response's `url` is the site's address: the
        root of its own host `<sitename>.<owner>.<base>/`, or, until the
        owner's certificate is ready, `<owner>.<base>/<sitename>/`. Use it as
        returned rather than composing it by hand, and take it from the latest
        response. A new site's access level is `only_me`.

        A new site's name becomes its address: lowercase letters, digits and
        hyphens, starting and ending with a letter or digit, at most
        63 characters, and not starting with `xn--`. Any other name is
        refused with `400 invalid_site_name`. A name equal to the address of
        one of the owner's older sites (for example `notes` beside an older
        `Notes`) is refused with `409 name_conflict`.
      tags: [Sites]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/sitename'
      requestBody:
        required: true
        content:
          application/gzip:
            schema:
              type: string
              format: binary
          application/zip:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Site created
          headers:
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteWithNote'
        '400':
          $ref: '#/components/responses/ManagementBadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Site already exists, or `name_conflict` (the name equals the address of one of the owner's older sites)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Compressed archive request exceeds 100 MiB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    put:
      summary: Update an existing site
      description: Upload a new gzip-compressed tar or ZIP version using the same archive safety, type, path, denylist, and size/count limits as site creation. Creates a new version number. Versions beyond the newest 5 (`QUOTA_MAX_VERSIONS`) are automatically cleaned up. `If-Match` is optional here, since only the owner deploys a site in their own namespace; team sites use the owner-qualified route, where it is required.
      tags: [Sites]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/sitename'
        - $ref: '#/components/parameters/publish'
        - name: If-Match
          in: header
          required: false
          description: Optional on this route; send the ETag captured before editing.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/gzip:
            schema:
              type: string
              format: binary
          application/zip:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Site updated
          headers:
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteWithNote'
        '400':
          $ref: '#/components/responses/ManagementBadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '413':
          description: Compressed archive request exceeds 100 MiB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    delete:
      summary: Delete a site
      description: The site stops serving at once and leaves every list. For 30 days it can be restored whole (`POST /api/sites/{sitename}/restore`) with its versions, saved data and history, access level, viewers and assets, and its name stays held (a create answers `409` `name_held`); then it is removed for good. Audited as `site_delete` with `restorable_until`.
      tags: [Sites]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/sitename'
      responses:
        '204':
          description: Site deleted
          headers:
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
        '400':
          $ref: '#/components/responses/SkillVersionRequired'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/collaboration/sites:
    get:
      summary: List owned and team sites
      description: |
        Returns canonical sites the authenticated user owns or that belong to
        a team the user is in. Legacy `GET /api/sites` remains owned-only for
        normal users. With `include=shared`, the sites the user, or a team
        they are in, is a named viewer of follow, as entries with
        `access_role: viewer` and `shared_via`; a viewer entry is a listing
        only, and no management route acts on it.
      tags: [Collaboration]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - name: include
          in: query
          required: false
          description: '`shared` adds the sites shared with the caller (Shared with me).'
          schema:
            type: string
            enum: [shared]
      responses:
        '200':
          description: Owned and team sites
          headers:
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollaborationSite'
        '400':
          $ref: '#/components/responses/SkillVersionRequired'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/collaboration/sites/{owner}/{sitename}:
    get:
      summary: Get an accessible canonical site
      description: Resolves the owner-qualified site and the caller's owner/member role.
      tags: [Collaboration]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/owner'
        - $ref: '#/components/parameters/sitename'
      responses:
        '200':
          description: Accessible site metadata
          headers:
            ETag:
              $ref: '#/components/headers/SiteETag'
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaborationSite'
        '400':
          $ref: '#/components/responses/SkillVersionRequired'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    put:
      summary: Deploy a new canonical-site version
      description: |
        Owner/member operation. Uploads a complete gzip-compressed TAR or ZIP
        artifact under the resource owner's existing namespace. Capture the
        ETag before editing and send it in `If-Match`; never refresh the ETag
        immediately before uploading a stale local snapshot. A stale validator
        returns 412 and must be reconciled deliberately.
      tags: [Collaboration]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/owner'
        - $ref: '#/components/parameters/sitename'
        - $ref: '#/components/parameters/publish'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/gzip:
            schema:
              type: string
              format: binary
          application/zip:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Version deployed
          headers:
            ETag:
              $ref: '#/components/headers/SiteETag'
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaborationSite'
        '400':
          $ref: '#/components/responses/ManagementBadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '413':
          description: Compressed archive request exceeds 100 MiB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/admin/export:
    get:
      summary: Export the audit or access log
      description: Admin only, browser session only. Streams every matching row as CSV (formula-safe) or NDJSON; audited as `admin_export` with the filters used. For `kind=audit` it takes the same filters as `/api/audit` (CSV ends with `actor_name`, `owner_name`, `site_name`); for `kind=access`, `owner` and `site`.
      tags: [Admin]
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum: [audit, access]
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, jsonl]
        - name: owner
          in: query
          schema:
            type: string
        - name: site
          in: query
          schema:
            type: string
        - name: actor
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        '200':
          description: The export, as an attachment
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/collaboration/sites/{owner}/{sitename}/export-link:
    post:
      summary: Get a download address for a whole site
      description: For the site's owner or a member of the owning team. Returns an address that works for 10 minutes with no key or cookie and downloads one zip of the site. Full-scope key or session.
      tags: [Sites]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/owner'
        - $ref: '#/components/parameters/sitename'
      responses:
        '200':
          description: The download address
          content:
            application/json:
              schema:
                type: object
                required: [url, expires_at]
                properties:
                  url:
                    type: string
                    description: GET it to download the zip; it carries its own credential, so share it with nobody.
                  expires_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/ManagementBadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/site-export/{token}:
    get:
      summary: Download a whole site
      description: The address `export-link` returns. One zip of `site.json`, `saved-data.json`, `saved-data-history.json`, `versions.json`, `assets.json`, the live version's files under `files/` and each uploaded file under `assets/<id>`. No key or cookie; the person who asked for the link must still own the site or be in its team, and the site must be the same one. Audited as `site_export`.
      tags: [Sites]
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The zip
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          description: The link is invalid or expired, or its site is gone or no longer the caller's.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/admin/users/{username}/export:
    get:
      summary: Export everything held about a disabled person
      description: Admin only, browser session only. Streams one zip - `account.json`, `teams.json`, `viewer-grants.json`, `api-keys.json`, `connected-apps.json`, `sessions.json` (metadata only, never a key, token or secret), `pending-grants.json` (viewer and team grants waiting for their provider-vouched email), `visits.jsonl` (the access-log rows of their own visits), `sites/<name>/` for every site they own including recently deleted ones (live files under `files/`, `site.json`, `saved-data.json`, `saved-data-history.json`, `versions.json`, `assets.json`, asset bytes under `assets/<id>`), and `audit-events.jsonl` (every audited action they took). Audited as `admin_user_export`.
      tags: [Admin]
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          description: The name is a team
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The person can still sign in; disable them first
...
components:
  headers:
    LatestSkillVersion:
      description: Latest skill bundle version published by this server.
      schema:
        type: string
        example: 0.8.1
    MinimumSkillVersion:
      description: Oldest skill version accepted for API-key management calls.
      schema:
        type: string
        example: 0.8.0
    SkillUpdateURL:
      description: Same-origin page describing the permissioned update flow.
      schema:
        type: string
        example: /install.html
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        An API key, for CI and other automation; agents normally connect
        through `/mcp` with the company sign-in instead. Each key has a scope
        chosen when it is created and an expiry (at most 365 days).
        `publish` (the default) deploys, updates and rolls back sites, reads
        versions, saved data and assets, and calls `/api/me` and `/mcp`.
        `full` adds everything else the person can do (deleting sites, access
        levels, viewers, teams, the audit and access logs, search), except
        administration. `offboard` (admins only) calls
        `POST /api/admin/users/disable` and nothing else. No key reaches any
        other `/api/admin/*` route. OAuth connector tokens work only on
        `/mcp`.
  parameters:
    SkillVersion:
      name: X-Skill-Version
      in: header
      required: false
      description: Required on API-key management requests made by an agent. Strict numeric major.minor.patch; the server rejects malformed or below-minimum values.
      schema:
        type: string
        pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)$'
      example: 0.8.1
    SimpleHostClient:
      name: X-Simple-Host-Client
      in: header
      required: false
      description: Intentional non-skill client classification. Direct API integrations send `api`; `control-ui` is reserved for the first-party account page. This is compatibility metadata, not authentication, and never replaces X-API-Key.
      schema:
        type: string
        enum: [api, control-ui]
      example: api
    sitename:
      name: sitename
      in: path
      required: true
      schema:
        type: string
      example: my-cool-site
...
    publish:
      name: publish
      in: query
      required: false
      description: 'Absent or `true`: the new version goes live. `false`: it is stored but not made live (the answer has `new_version`; `active_version` and the ETag stay), to be previewed and made live later by a rollback to it. Refused on a create (400 `publish_required`).'
      schema:
        type: boolean
...
  schemas:
...
    Site:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        active_version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        url:
          type: string
          description: The site's address. Use it as given.
          example: https://my-cool-site.alice-smith.simple-host.example.com/
        new_version:
          type: integer
          description: Only after a deploy with `publish=false`, the version it stored, which is not live.
        analytics:
          $ref: '#/components/schemas/SiteAnalytics'
...
    SiteWithNote:
      allOf:
        - $ref: '#/components/schemas/Site'
        - type: object
          properties:
            note:
              type: string
              description: Human-readable info about the site URL
              example: "Site is available at https://my-cool-site.alice-smith.simple-host.example.com/"
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        reason:
          type: string
          description: With `site_restricted_by_admin`, the admin's note.
...
  responses:
...
    SkillVersionRequired:
      description: A valid API key was supplied, but the management client omitted a supported skill version or explicit non-skill classification.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Simple-Host-Latest-Skill-Version:
          $ref: '#/components/headers/LatestSkillVersion'
        X-Simple-Host-Minimum-Skill-Version:
          $ref: '#/components/headers/MinimumSkillVersion'
        X-Simple-Host-Update-URL:
          $ref: '#/components/headers/SkillUpdateURL'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SkillVersionRequired'
    ManagementBadRequest:
      description: Invalid management request, or a valid API-key client that omitted a supported skill version or explicit non-skill classification.
      headers:
        Cache-Control:
          description: "no-store on skill-version bootstrap errors; otherwise handler-specific."
          schema:
            type: string
        X-Simple-Host-Latest-Skill-Version:
          $ref: '#/components/headers/LatestSkillVersion'
        X-Simple-Host-Minimum-Skill-Version:
          $ref: '#/components/headers/MinimumSkillVersion'
        X-Simple-Host-Update-URL:
          $ref: '#/components/headers/SkillUpdateURL'
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Error'
              - $ref: '#/components/schemas/SkillVersionRequired'
...
    Unauthorized:
      description: 'Missing or invalid credentials. For an `X-API-Key` that did not authenticate, `code` says why: `key_expired` (the body names the date; `expired_at` carries it), `key_revoked`, `key_owner_disabled` or `key_not_recognised`. A key with 14 days or less left gets `X-Key-Expires` (RFC 3339) and `X-Simple-Host-Notice` on every response.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    PreconditionFailed:
      description: The site incarnation or active version changed after the caller began editing
      headers:
        ETag:
          $ref: '#/components/headers/SiteETag'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SitePrecondition'
...
    TooManyRequests:
      description: Process-local rate limit exceeded; wait at least the advertised number of seconds before retrying
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'